backend/requirements.txt
# Django & REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# AI & ML Libraries
openai==1.10.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
nltk==3.8.1
transformers==4.36.2
torch==2.1.2

# Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
Pillow==10.2.0

# Development
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.27.0
